requirements.txt
pandas
numpy
matplotlib
scikit-learn
pyyaml
requests
nltk